1.21.11 (#50)
* Add current 1.21.11 update files

* Add breath of the nautilus effect

* Add nautilus armor item textures

* Fix scales by using floats in places where it is appropriate to use floats

* Use UnsafeKey where possible

* 25w45a/1.21.130.26

* Add IconTransformer to handle general UI icons (armor, tools, materials, smithing templates etc)

* Correctly handle missing vanilla textures in IconTransformer

* Add EnvironmentTransformer for the moon phase changes

* Address reviews

+++ converter/src/main/java/org/geysermc/pack/converter/type/texture/transformer/type/WeatherTransformer.java
@@ -58,7 +58,7 @@ public class WeatherTransformer implements TextureTransformer {
         BufferedImage snowImage = this.readImage(snowTexture);
         BufferedImage rainImage = this.readImage(rainTexture);
 
-        float factor = (float) snowImage.getWidth() / 64;
+        float factor = snowImage.getWidth() / 64f;
 
         BufferedImage weatherImage = new BufferedImage((int) (32 * factor), (int) (32 * factor), BufferedImage.TYPE_INT_ARGB);
